requirements.txt
streamlit
pandas
plotly
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-huggingface
langchain-pinecone
sentence-transformers
pinecone-client
python-dotenv
speechrecognition
folium
streamlit-folium